fix: match Figma for empty rules UI state
