+++ packages/security_center/lib/app_permissions/app_rules_page.dart
@@ -6,6 +6,7 @@ import 'package:security_center/app_permissions/snap_metadata_providers.dart';
 import 'package:security_center/app_permissions/snapd_interface.dart';
 import 'package:security_center/l10n.dart';
 import 'package:security_center/widgets/app_icon.dart';
+import 'package:security_center/widgets/empty_rules_tile.dart';
 import 'package:security_center/widgets/scrollable_page.dart';
 import 'package:security_center/widgets/tile_list.dart';
 import 'package:yaru/yaru.dart';
@@ -70,7 +71,7 @@ class _Body extends ConsumerWidget {
         if (rules.isEmpty) ...[
           const TileList(
             children: [
-              _EmptyRuleTile(),
+              EmptyRulesTile(),
             ],
           ),
           const SizedBox(height: 24),
@@ -166,20 +167,6 @@ class RuleTile extends StatelessWidget {
         icon: const Icon(YaruIcons.window_close),
         onPressed: onRemove,
       ),
-    );
-  }
-}
-
-class _EmptyRuleTile extends StatelessWidget {
-  const _EmptyRuleTile();
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      title: Center(
-        child: Text(AppLocalizations.of(context).snapRulesPageEmptyTileLabel),
-      ),
-      enabled: false,
     );
   }
 }
